# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Atomic Identity Habits API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /identities:
    get:
      operationId: listIdentities
      tags: [identities]
      summary: List all identities
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Identity"
    post:
      operationId: createIdentity
      tags: [identities]
      summary: Create an identity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
  /identities/{id}:
    patch:
      operationId: updateIdentity
      tags: [identities]
      summary: Update an identity
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Identity"
        "404":
          description: Not found
    delete:
      operationId: deleteIdentity
      tags: [identities]
      summary: Delete an identity
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── HABITS ─────────────────────────────────────────────────────────────────
...
  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List journal entries
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create or update a journal entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
  /journal/search:
    get:
      operationId: searchJournalEntries
      tags: [journal]
      summary: Search journal entries
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
  /journal/{id}:
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalEntryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Identity:
      type: object
      required: [id, userId, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        colorKey:
          type: ["string", "null"]
        habitCount:
          type: integer
        evidenceCount:
          type: integer
        consistencyScore:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    IdentityInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        colorKey:
          type: string
    IdentityUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        colorKey:
          type: string

    # ─── HABIT ───────────────────────────────────────────────────────────────
...
    JournalEntry:
      type: object
      required: [id, userId, date, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
        content:
          type: string
        mood:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    JournalEntryInput:
      type: object
      required: [date, content]
      properties:
        date:
          type: string
        content:
          type: string
        mood:
          type: string
        tags:
          type: array
          items:
            type: string
    JournalEntryUpdate:
      type: object
      properties:
        content:
          type: string
        mood:
          type: string
        tags:
          type: array
          items:
            type: string

    # ─── DASHBOARD ───────────────────────────────────────────────────────────
...